requirements.txt
python-telegram-bot==13.7
googletrans==4.0.0-rc1
configparser
speechrecognition
pydub